[loader]: Re-export symbols for C embedding, rename to `libjulia-internal`
We first rename `libjulia` to `libjulia-internal`, and `libjulialoader`
to `libjulia` so as to reduce confusion from the outside; `libjulia` is
the entrypoint for all usecases, and `libjulia-internal` just so happens
to be the chunk that contains all the actual code and linkage to other
libraries.

To support C linking against `libjulia` in a natural way, we must
re-export our symbols from `libjulia-internal`.  In order to do this we
create function trampolines from `libjulia` to `libjulia-internal` and
we define all symbols within `libjulia` to be filled in by the
initialization routines of `libjulia-internal`.

The function and data symbols are now maintained in lists stored in
`src/jl_exported_*.inc`.  These lists are consumed by `cli/Makefile`
where bindings are generated for them to be re-exported.  The function
trampolines are all naked functions that contain an architecture-
appropriate `jmp` instruction to `jl_*_addr` symbols, which are
initialized within `loader_export_symbols()` to point to the actual
function addresses within `libjulia-internal`.

The data symbols are merely defined as `void *` pointers within the
loader, then imported into `libjulia-internal` by the C linker.  This
backwards-dependency is slightly confusing, but it allows the symbols to
live within `libjulia` while being initialized by `libjulia-internal`.
At first glance it appears to be a circular dependency, however since
`libjulia` loads `libjulia-internal` through `dlopen()` first, then
`libjulia-internal` finds `libjulia` already loaded, the circular
dependency is a non-issue.  C code that wishes to link against
`libjulia` should do so directly, and the values of all symbols, while
initially being NULL, will be set to their correct values after
initialization through `load_libjulia_internal()` is complete.